Renamed this package from testing to refactor to reduce confusion with testing directory
