+++ src/main/java/refactor/ProfileEntity.java
@@ -1,4 +1,4 @@
-package test;
+package refactor;
 
 import javax.persistence.*;
