+++ src/main/java/refactor/SessionFactoryProvider.java
@@ -1,4 +1,4 @@
-package test;
+package refactor;
 
 import org.hibernate.SessionFactory;
 import org.hibernate.boot.Metadata;
